requirements.txt
# Core packages (binary distributions)
numpy==2.2.2
pandas==2.2.3

# AWS and utilities
boto3==1.26.137
python-dotenv==1.0.0
botocore==1.29.137

# Web and API
Flask==2.2.5
flask-cors==4.0.0
requests==2.31.0
websocket-client==1.6.3

# Other utilities
python-dateutil==2.8.2
werkzeug==2.2.3
typing-extensions==4.7.1
flasgger==0.9.7

# Airflow and ETL
apache-airflow==2.7.1
apache-airflow-providers-snowflake==2.0.1
apache-airflow-providers-amazon==3.0.0
apache-airflow-providers-apache-kafka==2.0.0
snowflake-connector-python==2.7.9
kafka-python==2.0.2
